requirements.txt
streamlit
langgraph
langchain-groq
langchain_google_community
google-auth-oauthlib
duckduckgo-search
streamlit-mic-recorder
gTTS
chromadb
sentence-transformers
python-dotenv
beautifulsoup4